+++ ZuneCraft/src/Core/Service.cpp
@@ -4,7 +4,7 @@ namespace ZuneCraft {
 	std::vector<IService*> Services::s_Services;
 	
 	void Services::Shutdown() {
-		for (int i = 0; i < s_Services.size(); i++) {
+		for (size_t i = 0; i < s_Services.size(); i++) {
 			delete s_Services[i];
 		}
 	}
